tkesar_wk1_2ndsubmission
Jin and Rocky, I tried to use some (not all) elements of week 1 tasks to open, manipulate, and plot a csv file with my data. Its still very inefficient. I was wondering if you could look at it and suggest ways to improve the code. Multiple options / ways of doing the simple tasks would greatly help me learn. 
I also want to make a plot where individual subjects are connected by lines, and group means are also shown on the same lineplot. But this is what I have for now and wanted to send your way for feedback (please). Thanks in advance.