Add files via upload
I am trying to open multiple csv files all of which have the same format as the one attached. I ran into many problems trying to concatenate them (which is the goal), but I wanted to share here so that you can intervene early in case I am going down the wrong path. For code testing purposes it'd be ok to make 4 copies of the same csv file and assign them df1, df2, df3, df4, etc. Thanks in advance!